requirements.txt
python-3.6.9
flask-cors
numpy
newspaper3k
sklearn
pandas
flask
nltk
gunicorn